requirements.txt
discord.py
requests[socks]